+++ packages/deepsea-tools/src/utils/createRequestFn.ts
@@ -1,3 +1,5 @@
+import { $ZodType, parse } from "zod/v4/core"
+
 import { assignFnName } from "./assignFnName"
 import { createFnWithMiddleware, Middleware } from "./createFnWithMiddleware"
 
@@ -12,7 +14,7 @@ export interface ResponseData<Data = unknown> {
 
 export type WithPromise<T> = T | Promise<T>
 
-export type OriginalRequestFn = (...args: any[]) => WithPromise<ResponseData<any> | undefined | void>
+export type OriginalRequestFn<Params extends any[] = any[]> = (...args: Params) => WithPromise<ResponseData<any> | undefined | void>
 
 export type RequestFnReturnType<Fn extends OriginalRequestFn> = Awaited<ReturnType<Fn>> extends ResponseData<infer T> ? T : undefined
 
@@ -30,25 +32,41 @@ export function isRequestFn(fn: unknown): fn is RequestFn<OriginalRequestFn> {
 
 const globalRequestFnMiddlewares: Middleware[] = []
 
-export function createRequestFn<Fn extends OriginalRequestFn>(fn: Fn): RequestFn<Fn> {
+export interface CreateResponseFnParams<Params extends [arg?: unknown]> {
+    fn: OriginalRequestFn<Params>
+    schema?: Params extends [] ? undefined : $ZodType<Params[0]>
+    name?: string
+}
+
+export function createRequestFn<Fn extends OriginalRequestFn>(fn: Fn): RequestFn<Fn>
+export function createRequestFn<Params extends [arg?: unknown]>(params: CreateResponseFnParams<Params>): RequestFn<OriginalRequestFn<Params>>
+export function createRequestFn(fnOrParams: any): any {
+    const { fn, schema, name } = (typeof fnOrParams === "function" ? { fn: fnOrParams } : fnOrParams) as CreateResponseFnParams<[arg: any]>
+
+    type Fn = typeof fn
+
     async function request(...args: Parameters<Fn>): Promise<RequestFnReturnType<Fn>> {
+        if (schema) args = [parse(schema, args.at(0))]
         const result = await fn(...args)
         if (result === undefined) return undefined as RequestFnReturnType<Fn>
         if (!result.success) throw result.message instanceof Error ? result.message : new Error(result.message)
         return result.data
     }
 
-    assignFnName(request, fn)
+    assignFnName(request, name ?? fn)
+
     const newRequest = createFnWithMiddleware(request, { global: globalRequestFnMiddlewares }) as RequestFn<Fn>
+
     Object.defineProperty(newRequest, requestFnSymbol, { value: true })
     newRequest.preset = function preset(...args: Parameters<Fn>): () => Promise<RequestFnReturnType<Fn>> {
         async function requestWithPreset() {
             return await newRequest(...args)
         }
 
-        assignFnName(requestWithPreset, fn)
+        assignFnName(requestWithPreset, name ?? fn)
         return requestWithPreset
     }
+
     return newRequest
 }